Fix minor warning

+++ src/workflow/oph_workflow_setup.c
@@ -249,7 +249,7 @@ int oph_workflow_validate(oph_workflow * workflow)
 
 	// Create graph from tasks
 	workflow_node *graph = NULL;
-	graph = (workflow_node *) calloc(workflow->tasks_num, sizeof(workflow_node));
+	graph = (workflow_node *) calloc((unsigned int)workflow->tasks_num, sizeof(workflow_node));
 	if (!graph)
 		return OPH_WORKFLOW_EXIT_MEMORY_ERROR;
